Fix button shifting due to screen transition animations (#2)
* fix: :bug: Hacky fix for buttons shifting due to animation when switching between screens

The fix is essentially preventing the root position from being set to the animated position. We check if we have animated and if our position is not the expected calculated root. If this is the case, we then will move back by the delta and thus cancel out the problem.

* fix: :bug: Hacky fix for buttons shifting due to animation when switching between screens

The fix is essentially preventing the root position from being set to the animated position. We check if we have animated and if our position is not the expected calculated root. If this is the case, we then will move back by the delta and thus cancel out the problem.